requirements.txt
flask>=2.3.0
requests>=2.30.0
python-dotenv>=1.0.1